@@ -6,7 +6,7 @@
     "browser": "webpack --config browser/webpack.config.js && node browser/scripts/benchmark.js"
   },
   "dependencies": {
-    "@emotion/react": "^11.4.1",
+    "@emotion/react": "^11.5.0",
     "@emotion/styled": "^11.3.0",
     "@material-ui/core": "^5.0.0-beta.4",
     "@material-ui/icons": "^5.0.0-beta.4",
@@ -19,14 +19,14 @@
     "css-loader": "^6.4.0",
     "html-webpack-plugin": "^5.4.0",
     "mui-plus": "^0.0.15",
-    "playwright": "^1.15.2",
+    "playwright": "^1.16.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-spring": "^9.2.4",
     "react-virtualized": "^9.22.3",
     "serve-handler": "^6.1.3",
-    "style-loader": "^3.3.0",
-    "webpack": "^5.58.2",
-    "webpack-cli": "^4.9.0"
+    "style-loader": "^3.3.1",
+    "webpack": "^5.60.0",
+    "webpack-cli": "^4.9.1"
   }
 }
