@@ -6,7 +6,7 @@
     "benchmark:browser": "yarn workspace benchmark browser",
     "docs:dev": "yarn workspace docs dev",
     "docs:api": "yarn docs:api:build",
-    "docs:api:build": "cross-env BABEL_ENV=development babel-node -x .ts,.tsx,.js ./docs/scripts/buildApi.ts ./docs/pages/api-docs/data-grid",
+    "docs:api:build": "cross-env BABEL_ENV=development babel-node -i '/node_modules/(?!@material-ui)/' -x .ts,.tsx,.js ./docs/scripts/buildApi.ts ./docs/pages/api-docs/data-grid",
     "docs:build": "yarn workspace docs build",
     "docs:export": "yarn workspace docs export",
     "docs:typescript:formatted": "yarn workspace docs typescript:transpile",
@@ -57,32 +57,33 @@
     "@babel/plugin-transform-object-assign": "^7.14.5",
     "@babel/plugin-transform-runtime": "^7.15.8",
     "@babel/register": "^7.15.3",
-    "@emotion/cache": "^11.4.0",
-    "@emotion/react": "^11.4.1",
+    "@babel/runtime": "^7.15.4",
+    "@emotion/cache": "^11.5.0",
+    "@emotion/react": "^11.5.0",
     "@emotion/styled": "^11.3.0",
     "@eps1lon/enzyme-adapter-react-17": "^0.1.0",
-    "@material-ui/monorepo": "https://github.com/mui-org/material-ui.git#master",
+    "@material-ui/monorepo": "https://github.com/m4theushw/material-ui-fork.git#master",
     "@mui/icons-material": "^5.0.4",
     "@mui/material": "^5.0.4",
     "@mui/styles": "^5.0.1",
     "@mui/utils": "^5.0.1",
     "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-node-resolve": "^13.0.5",
+    "@rollup/plugin-node-resolve": "^13.0.6",
     "@rollup/plugin-replace": "^3.0.0",
     "@testing-library/react": "^12.1.2",
     "@types/chai": "^4.2.22",
     "@types/chai-dom": "^0.0.11",
-    "@types/enzyme": "^3.10.9",
+    "@types/enzyme": "^3.10.10",
     "@types/mocha": "^9.0.0",
-    "@types/node": "^16.11.1",
+    "@types/node": "^16.11.4",
     "@types/prettier": "^2.4.1",
     "@types/react": "^17.0.30",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^17.0.10",
     "@types/requestidlecallback": "^0.3.4",
-    "@types/sinon": "^10.0.4",
+    "@types/sinon": "^10.0.5",
     "@types/yargs": "^17.0.4",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.2.0",
+    "@typescript-eslint/parser": "^5.2.0",
     "argos-cli": "^0.3.3",
     "babel-plugin-module-resolver": "^4.1.0",
     "babel-plugin-optimize-clsx": "^2.6.2",
@@ -97,9 +98,9 @@
     "eslint": "^7.32.0",
     "eslint-config-airbnb-typescript": "^12.3.1",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-webpack": "^0.13.1",
+    "eslint-import-resolver-webpack": "^0.13.2",
     "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-jsdoc": "^36.1.1",
+    "eslint-plugin-jsdoc": "^37.0.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-mocha": "^9.0.0",
     "eslint-plugin-prettier": "^3.4.1",
@@ -111,7 +112,7 @@
     "jss": "^10.8.0",
     "jss-plugin-template": "^10.8.1",
     "jss-rtl": "^0.3.0",
-    "karma": "^6.3.4",
+    "karma": "^6.3.6",
     "karma-browserstack-launcher": "^1.6.0",
     "karma-chrome-launcher": "^3.1.0",
     "karma-mocha": "^2.0.1",
@@ -120,7 +121,7 @@
     "lerna": "^4.0.0",
     "mocha": "^9.1.3",
     "nyc": "^15.1.0",
-    "playwright": "^1.15.2",
+    "playwright": "^1.16.1",
     "prettier": "^2.4.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -138,10 +139,10 @@
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
     "stylelint-processor-styled-components": "^1.10.0",
-    "typedoc": "^0.22.5",
+    "typedoc": "^0.22.7",
     "typescript": "^4.4.4",
-    "webpack": "^5.58.2",
-    "webpack-cli": "^4.9.0",
+    "webpack": "^5.60.0",
+    "webpack-cli": "^4.9.1",
     "yargs": "^17.2.1",
     "yarn-deduplicate": "^3.1.0"
   },
