@@ -2,7 +2,7 @@ import * as React from 'react';
 import { useGridSelector } from '../hooks/utils/useGridSelector';
 import { gridRowCountSelector } from '../hooks/features/rows/gridRowsSelector';
 import { selectedGridRowsCountSelector } from '../hooks/features/selection/gridSelectionSelector';
-import { visibleGridRowCountSelector } from '../hooks/features/filter/gridFilterSelector';
+import { gridVisibleRowCountSelector } from '../hooks/features/filter/gridFilterSelector';
 import { useGridApiContext } from '../hooks/utils/useGridApiContext';
 import { GridRowCount } from './GridRowCount';
 import { GridSelectedRowCount } from './GridSelectedRowCount';
@@ -15,7 +15,7 @@ export const GridFooter = React.forwardRef<HTMLDivElement, GridFooterContainerPr
     const rootProps = useGridRootProps();
     const totalRowCount = useGridSelector(apiRef, gridRowCountSelector);
     const selectedRowCount = useGridSelector(apiRef, selectedGridRowsCountSelector);
-    const visibleRowCount = useGridSelector(apiRef, visibleGridRowCountSelector);
+    const visibleRowCount = useGridSelector(apiRef, gridVisibleRowCountSelector);
 
     const selectedRowCountElement =
       !rootProps.hideFooterSelectedRowCount && selectedRowCount > 0 ? (
