@@ -248,11 +248,11 @@ export interface GridClasses {
   /**
    * Styles applied to the left scroll area element.
    */
-  scrollAreaLeft: string;
+  'scrollArea--left': string;
   /**
    * Styles applied to the right scroll area element.
    */
-  scrollAreaRight: string;
+  'scrollArea--right': string;
   /**
    * Styles applied to the footer selected row count element.
    */
@@ -269,14 +269,6 @@ export interface GridClasses {
    * Styles applied to the toolbar filter list element.
    */
   toolbarFilterList: string;
-  /**
-   * Styles applied to the viewport element.
-   */
-  viewport: string;
-  /**
-   * Styles applied to the window element.
-   */
-  window: string;
   /**
    * Styles applied to both the cell and the column header if `showColumnRightBorder={true}`.
    */
