@@ -13,16 +13,14 @@ import { useGridLogger } from '../../utils/useGridLogger';
 import { filterableGridColumnsIdsSelector } from '../columns/gridColumnsSelector';
 import { useGridState } from '../../utils/useGridState';
 import { GridPreferencePanelsValue } from '../preferencesPanel/gridPreferencePanelsValue';
-import { sortedGridRowsSelector } from '../sorting/gridSortingSelector';
+import { gridSortedRowIdsSelector } from '../sorting/gridSortingSelector';
 import { getDefaultGridFilterModel } from './gridFilterState';
 import { GridFilterModel } from '../../../models/gridFilterModel';
-import {
-  gridVisibleRowsLookupSelector,
-  visibleSortedGridRowsSelector,
-  gridFilterModelSelector,
-} from './gridFilterSelector';
+import { gridVisibleSortedRowEntriesSelector, gridFilterModelSelector } from './gridFilterSelector';
 import { useGridStateInit } from '../../utils/useGridStateInit';
 import { useFirstRender } from '../../utils/useFirstRender';
+
+type GridFilterItemApplier = (rowId: GridRowId) => boolean;
 
 const checkFilterModelValidity = (model: GridFilterModel) => {
   if (model.items.length > 1) {
@@ -74,7 +72,7 @@ export const useGridFilter = (
 
   const [, setGridState, forceUpdate] = useGridState(apiRef);
 
-  apiRef.current.updateControlState({
+  apiRef.current.unsafe_updateControlState({
     stateId: 'filter',
     propModel: props.filterModel,
     propOnChange: props.onFilterModelChange,
@@ -82,109 +80,124 @@ export const useGridFilter = (
     changeEvent: GridEvents.filterModelChange,
   });
 
-  const applyFilter = React.useCallback(
-    (filterItem: GridFilterItem, linkOperator: GridLinkOperator = GridLinkOperator.And) => {
-      if (!filterItem.columnField || !filterItem.operatorValue) {
-        return;
+  const buildAggregatedFilterApplier = React.useCallback(
+    (filterModel: GridFilterModel): GridFilterItemApplier | null => {
+      const { items, linkOperator = GridLinkOperator.And } = filterModel;
+
+      const getFilterCallbackFromItem = (
+        filterItem: GridFilterItem,
+      ): GridFilterItemApplier | null => {
+        if (!filterItem.columnField || !filterItem.operatorValue) {
+          return null;
+        }
+
+        const column = apiRef.current.getColumn(filterItem.columnField);
+        if (!column) {
+          return null;
+        }
+
+        const parsedValue = column.valueParser
+          ? column.valueParser(filterItem.value)
+          : filterItem.value;
+        const newFilterItem: GridFilterItem = { ...filterItem, value: parsedValue };
+
+        const filterOperators = column.filterOperators;
+        if (!filterOperators?.length) {
+          throw new Error(`MUI: No filter operators found for column '${column.field}'.`);
+        }
+
+        const filterOperator = filterOperators.find(
+          (operator) => operator.value === newFilterItem.operatorValue,
+        )!;
+        if (!filterOperator) {
+          throw new Error(
+            `MUI: No filter operator found for column '${column.field}' and operator value '${newFilterItem.operatorValue}'.`,
+          );
+        }
+
+        const applyFilterOnRow = filterOperator.getApplyFilterFn(newFilterItem, column)!;
+        if (typeof applyFilterOnRow !== 'function') {
+          return null;
+        }
+
+        return (rowId: GridRowId) => {
+          const cellParams = apiRef.current.getCellParams(rowId, newFilterItem.columnField!);
+
+          return applyFilterOnRow(cellParams);
+        };
+      };
+
+      const appliers = items
+        .map(getFilterCallbackFromItem)
+        .filter((callback): callback is GridFilterItemApplier => !!callback);
+
+      if (appliers.length === 0) {
+        return null;
       }
 
-      const column = apiRef.current.getColumn(filterItem.columnField);
+      return (rowId: GridRowId) => {
+        // Return `false` as soon as we have a failing filter
+        if (linkOperator === GridLinkOperator.And) {
+          return appliers.every((applier) => applier(rowId));
+        }
 
-      if (!column) {
-        return;
-      }
+        // Return `true` as soon as we have a passing filter
+        return appliers.some((applier) => applier(rowId));
+      };
+    },
+    [apiRef],
+  );
 
-      const parsedValue = column.valueParser
-        ? column.valueParser(filterItem.value)
-        : filterItem.value;
-      const newFilterItem = { ...filterItem, value: parsedValue };
+  const applyFilters = React.useCallback<GridFilterApi['unsafe_applyFilters']>(() => {
+    setGridState((state) => {
+      const filterModel = gridFilterModelSelector(state);
 
-      logger.debug(
-        `Filtering column: ${newFilterItem.columnField} ${newFilterItem.operatorValue} ${newFilterItem.value} `,
-      );
-
-      const filterOperators = column.filterOperators;
-      if (!filterOperators?.length) {
-        throw new Error(`MUI: No filter operators found for column '${column.field}'.`);
-      }
-
-      const filterOperator = filterOperators.find(
-        (operator) => operator.value === newFilterItem.operatorValue,
-      )!;
-      if (!filterOperator) {
-        throw new Error(
-          `MUI: No filter operator found for column '${column.field}' and operator value '${newFilterItem.operatorValue}'.`,
-        );
-      }
-
-      const applyFilterOnRow = filterOperator.getApplyFilterFn(newFilterItem, column)!;
-      if (typeof applyFilterOnRow !== 'function') {
-        return;
-      }
-
-      setGridState((state) => {
-        const visibleRowsLookup = { ...gridVisibleRowsLookupSelector(state) };
-
-        // We run the selector on the state here to avoid rendering the rows and then filtering again.
-        // This way we have latest rows on the first rendering
-        const rows = sortedGridRowsSelector(state);
-
-        rows.forEach((row: GridRowModel, id: GridRowId) => {
-          const params = apiRef.current.getCellParams(id, newFilterItem.columnField!);
-
-          const isShown = applyFilterOnRow(params);
-          if (visibleRowsLookup[id] == null) {
-            visibleRowsLookup[id] = isShown;
-          } else {
-            visibleRowsLookup[id] =
-              linkOperator === GridLinkOperator.And
-                ? visibleRowsLookup[id] && isShown
-                : visibleRowsLookup[id] || isShown;
-          }
-        });
-
+      if (props.filterMode === GridFeatureModeConstant.server) {
         return {
           ...state,
           filter: {
             ...state.filter,
-            visibleRowsLookup,
-            visibleRows: Object.entries(visibleRowsLookup)
-              .filter(([, isVisible]) => isVisible)
-              .map(([id]) => id),
+            visibleRowsLookup: {},
+            visibleRows: null,
           },
         };
+      }
+
+      // No filter to apply
+      const filteringMethod = buildAggregatedFilterApplier(filterModel);
+      if (!filteringMethod) {
+        return {
+          ...state,
+          filter: {
+            ...state.filter,
+            visibleRowsLookup: {},
+            visibleRows: null,
+          },
+        };
+      }
+
+      const rowIds = gridSortedRowIdsSelector(apiRef.current.state);
+      const visibleRowsLookup: Record<GridRowId, boolean> = {};
+      rowIds.forEach((rowId) => {
+        visibleRowsLookup[rowId] = filteringMethod(rowId);
       });
-      forceUpdate();
-    },
-    [apiRef, forceUpdate, logger, setGridState],
-  );
 
-  const applyFilters = React.useCallback<GridFilterApi['applyFilters']>(() => {
-    if (props.filterMode === GridFeatureModeConstant.server) {
-      forceUpdate();
-      return;
-    }
-
-    // Clearing filtered rows
-    setGridState((state) => ({
-      ...state,
-      filter: {
-        ...state.filter,
-        visibleRowsLookup: {},
-        visibleRows: null,
-      },
-    }));
-
-    const { items, linkOperator } = gridFilterModelSelector(apiRef.current.state);
-
-    items.forEach((filterItem) => {
-      apiRef.current.applyFilter(filterItem, linkOperator);
+      return {
+        ...state,
+        filter: {
+          ...state.filter,
+          visibleRowsLookup,
+          visibleRows: Object.entries(visibleRowsLookup)
+            .filter(([, isVisible]) => isVisible)
+            .map(([id]) => id),
+        },
+      };
     });
     apiRef.current.publishEvent(GridEvents.visibleRowsSet);
     forceUpdate();
-  }, [apiRef, setGridState, forceUpdate, props.filterMode]);
+  }, [apiRef, setGridState, forceUpdate, props.filterMode, buildAggregatedFilterApplier]);
 
-  const upsertFilter = React.useCallback<GridFilterApi['upsertFilter']>(
+  const upsertFilterItem = React.useCallback<GridFilterApi['upsertFilterItem']>(
     (item) => {
       logger.debug('Upserting filter');
 
@@ -224,12 +237,12 @@ export const useGridFilter = (
           filter: { ...state.filter, filterModel: { ...state.filter.filterModel, items } },
         };
       });
-      apiRef.current.applyFilters();
+      apiRef.current.unsafe_applyFilters();
     },
     [apiRef, logger, setGridState, props.disableMultipleColumnsFiltering],
   );
 
-  const deleteFilter = React.useCallback<GridFilterApi['deleteFilter']>(
+  const deleteFilterItem = React.useCallback<GridFilterApi['deleteFilterItem']>(
     (item) => {
       logger.debug(`Deleting filter on column ${item.columnField} with value ${item.value}`);
       setGridState((state) => {
@@ -243,9 +256,9 @@ export const useGridFilter = (
         };
       });
       if (gridFilterModelSelector(apiRef.current.state).items.length === 0) {
-        apiRef.current.upsertFilter({});
+        apiRef.current.upsertFilterItem({});
       }
-      apiRef.current.applyFilters();
+      apiRef.current.unsafe_applyFilters();
     },
     [apiRef, logger, setGridState],
   );
@@ -259,7 +272,7 @@ export const useGridFilter = (
         const lastFilter =
           filterModel.items.length > 0 ? filterModel.items[filterModel.items.length - 1] : null;
         if (!lastFilter || lastFilter.columnField !== targetColumnField) {
-          apiRef.current.upsertFilter({ columnField: targetColumnField });
+          apiRef.current.upsertFilterItem({ columnField: targetColumnField });
         }
       }
       apiRef.current.showPreferences(GridPreferencePanelsValue.filters);
@@ -272,14 +285,14 @@ export const useGridFilter = (
     apiRef.current.hidePreferences();
   }, [apiRef, logger]);
 
-  const applyFilterLinkOperator = React.useCallback<GridFilterApi['applyFilterLinkOperator']>(
+  const setFilterLinkOperator = React.useCallback<GridFilterApi['setFilterLinkOperator']>(
     (linkOperator) => {
       logger.debug('Applying filter link operator');
       setGridState((state) => ({
         ...state,
         filter: { ...state.filter, filterModel: { ...state.filter.filterModel, linkOperator } },
       }));
-      apiRef.current.applyFilters();
+      apiRef.current.unsafe_applyFilters();
     },
     [apiRef, logger, setGridState],
   );
@@ -298,25 +311,24 @@ export const useGridFilter = (
             filterModel: model,
           },
         }));
-        apiRef.current.applyFilters();
+        apiRef.current.unsafe_applyFilters();
       }
     },
     [apiRef, logger, setGridState],
   );
 
-  const getVisibleRowModels = React.useCallback<GridFilterApi['getVisibleRowModels']>(
-    () => visibleSortedGridRowsSelector(apiRef.current.state),
-    [apiRef],
-  );
+  const getVisibleRowModels = React.useCallback<GridFilterApi['getVisibleRowModels']>(() => {
+    const visibleSortedRows = gridVisibleSortedRowEntriesSelector(apiRef.current.state);
+    return new Map<GridRowId, GridRowModel>(visibleSortedRows.map((row) => [row.id, row.model]));
+  }, [apiRef]);
 
   useGridApiMethod<GridFilterApi>(
     apiRef,
     {
-      applyFilterLinkOperator,
-      applyFilters,
-      applyFilter,
-      deleteFilter,
-      upsertFilter,
+      setFilterLinkOperator,
+      unsafe_applyFilters: applyFilters,
+      deleteFilterItem,
+      upsertFilterItem,
       setFilterModel,
       showFilterPanel,
       hideFilterPanel,
@@ -333,10 +345,10 @@ export const useGridFilter = (
 
     filterModel.items.forEach((filter) => {
       if (!columnsIds.find((field) => field === filter.columnField)) {
-        apiRef.current.deleteFilter(filter);
+        apiRef.current.deleteFilterItem(filter);
       }
     });
-    apiRef.current.applyFilters();
+    apiRef.current.unsafe_applyFilters();
   }, [apiRef, logger]);
 
   React.useEffect(() => {
@@ -345,8 +357,8 @@ export const useGridFilter = (
     }
   }, [apiRef, logger, props.filterModel]);
 
-  useFirstRender(() => apiRef.current.applyFilters());
+  useFirstRender(() => apiRef.current.unsafe_applyFilters());
 
-  useGridApiEventHandler(apiRef, GridEvents.rowsSet, apiRef.current.applyFilters);
+  useGridApiEventHandler(apiRef, GridEvents.rowsSet, apiRef.current.unsafe_applyFilters);
   useGridApiEventHandler(apiRef, GridEvents.columnsChange, onColUpdated);
 };
